category/categoryMapper/categoryService 타입 변경으로 인한 수정

+++ src/main/java/com/example/demo/category/CategoryMapper.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.query.Param;
 @Mapper
 public interface CategoryMapper {
 
-	String findCategoryNameById(@Param("categoryID") Long categoryID);
+	String findCategoryNameById(@Param("categoryID") Integer categoryID);
 
 }
